backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==21.2.0
supabase==2.4.6
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
celery[redis]==5.3.6
redis==5.0.4
h3==3.7.7
haversine==2.8.1
shapely==2.0.4
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
opencv-python-headless==4.9.0.80
ultralytics==8.2.0
hachoir==3.2.0
ffmpeg-python==0.2.0
httpx==0.27.0
tenacity==8.3.0
razorpay==1.4.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
cryptography==42.0.7
pydantic==2.7.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
structlog==24.1.0
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk==2.2.0
ruff==0.4.2
mypy==1.10.0
pre-commit==3.7.0
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
factory-boy==3.3.0
freezegun==1.4.0
pytz==2024.1
